Give placement its own module
